Move registry properties base classes to properties package
