+++ micrometer-spring-legacy/src/main/java/io/micrometer/spring/autoconfigure/export/properties/PropertiesConfigAdapter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.spring.autoconfigure.export;
+package io.micrometer.spring.autoconfigure.export.properties;
 
 import org.springframework.util.Assert;
